+++ data_adapter/collection.py
@@ -4,7 +4,9 @@ from dataclasses import dataclass
 from enum import IntEnum
 from typing import List, Optional, Union
 
-from data_adapter import core, settings
+import pandas
+
+from data_adapter import core, ontology, settings
 
 
 class CollectionError(Exception):
@@ -22,12 +24,24 @@ class Artifact:
     group: str
     artifact: str
     version: str
-    filename: str
-    subject: Optional[str]
-    datatype: DataType
+    filename: Optional[str] = None
+    datatype: DataType = DataType.Scalar
+    multiple_types: bool = False
 
-    def path(self):
+    @property
+    def path(self) -> pathlib.Path:
         return settings.COLLECTIONS_DIR / self.collection / self.group / self.artifact / self.version
+
+    @property
+    def metadata(self) -> dict:
+        with open(self.path / self.get_filename(".json"), "r", encoding="utf-8") as metadata_file:
+            return json.load(metadata_file)
+
+    def get_filename(self, suffix: str) -> str:
+        for file in self.path.iterdir():
+            if file.suffix == suffix:
+                return file.name
+        raise FileNotFoundError(f"Artifact file with {suffix=} not found.")
 
 
 def check_collection_meta(collection_meta: dict):
@@ -49,7 +63,7 @@ def check_collection_meta(collection_meta: dict):
     # Check if artifact info keys are missing:
     for group, artifacts in collection_meta["artifacts"].items():
         for artifact, artifact_infos in artifacts.items():
-            for key in ("latest_version", "subject", "datatype"):
+            for key in ("latest_version",):
                 if key not in artifact_infos:
                     raise CollectionError(
                         f"Collection metadata is invalid ({group=}, {artifact=} misses {key=}). "
@@ -57,22 +71,54 @@ def check_collection_meta(collection_meta: dict):
                     )
 
 
-def get_metadata_from_artifact(artifact: Artifact) -> dict:
+def infer_collection_metadata(collection: str, collection_meta: dict) -> dict:
     """
-    Returns metadata from given artifact.
+    Interferes downloaded collection and updates names and subjects of artifacts in collection metadata file
 
     Parameters
     ----------
-    artifact: Artifact
-        Artifact to get metadata from
+    collection: str
+        Name of collection
+    collection_meta : dict
+        Metadata of collection to be updated
 
     Returns
     -------
     dict
-        Metadata from artifact
+        Updated collection metadata
     """
-    with open(artifact.path() / f"{artifact.filename}.json", "r", encoding="utf-8") as metadata_file:
-        return json.load(metadata_file)
+    for group_name, artifacts in collection_meta["artifacts"].items():
+        for artifact_name in artifacts:
+            version = collection_meta["artifacts"][group_name][artifact_name]["latest_version"]
+
+            artifact = Artifact(collection, group_name, artifact_name, version)
+            metadata = artifact.metadata
+
+            # Check if artifact contains multiple (sub-)processes
+            try:
+                type_field = [
+                    field for field in metadata["resources"][0]["schema"]["fields"] if field["name"] == "type"
+                ][0]
+            except IndexError:
+                type_field = None
+
+            if type_field:
+                collection_meta["artifacts"][group_name][artifact_name]["multiple_types"] = True
+                collection_meta["artifacts"][group_name][artifact_name]["names"] = get_subprocesses_from_artifact(
+                    artifact
+                )
+                collection_meta["artifacts"][group_name][artifact_name]["subjects"] = [
+                    ontology.get_name_from_annotation(value_reference)
+                    for value_reference in type_field["valueReference"]
+                ]
+            else:
+                collection_meta["artifacts"][group_name][artifact_name]["multiple_types"] = False
+                collection_meta["artifacts"][group_name][artifact_name]["names"] = [metadata["name"]]
+                collection_meta["artifacts"][group_name][artifact_name]["subjects"] = [ontology.get_subject(metadata)]
+
+            collection_meta["artifacts"][group_name][artifact_name]["datatype"] = get_data_type(metadata)
+
+    return collection_meta
 
 
 def get_data_type(metadata: Union[str, pathlib.Path, dict]):
@@ -81,6 +127,25 @@ def get_data_type(metadata: Union[str, pathlib.Path, dict]):
         if field["name"].startswith("timeindex"):
             return DataType.Timeseries
     return DataType.Scalar
+
+
+def get_subprocesses_from_artifact(artifact: Artifact) -> List[str]:
+    """
+    Return list of subprocesses for given artifact
+
+    Returns entries of column "type" as list.
+
+    Parameters
+    ----------
+    artifact: Artifact
+        Read type column of given artifact
+
+    Returns
+    -------
+    List[str]
+        List of subprocesses in given artifact
+    """
+    return list(pandas.read_csv(artifact.path / artifact.get_filename(".csv"), usecols=("type",))["type"])
 
 
 def get_collection_meta(collection: str) -> dict:
@@ -134,8 +199,8 @@ def get_artifacts_from_collection(collection: str, process: Optional[str] = None
     artifacts = []
     for group in collection_meta["artifacts"]:
         for artifact, artifact_info in collection_meta["artifacts"][group].items():
-            process_name = artifact_info["subject"] if settings.USE_ANNOTATIONS else artifact_info["name"]
-            if process and process_name != process:
+            process_names = artifact_info["subjects"] if settings.USE_ANNOTATIONS else artifact_info["names"]
+            if process and process not in process_names:
                 continue
             filename = artifact
             artifacts.append(
@@ -145,8 +210,8 @@ def get_artifacts_from_collection(collection: str, process: Optional[str] = None
                     artifact,
                     artifact_info["latest_version"],
                     filename,
-                    subject=process,
                     datatype=DataType(artifact_info["datatype"]),
+                    multiple_types=artifact_info["multiple_types"],
                 )
             )
     return artifacts